+++ tests/test_numpy.py
@@ -14,25 +14,25 @@ else:
 from . import MeasureItTestCase, math_is_hard
 
 from measure_it import measure_iter
-from measure_it.numpy import StatsMetric, PlotMetric
+from measure_it.numpy import TableMetric, PlotMetric
 
-class StatsMetricTestCase(MeasureItTestCase):
+class TableMetricTestCase(MeasureItTestCase):
 
     def test_stats(self):
-        StatsMetric.dump_atexit = False
-        StatsMetric.outfile = StringIO()
+        TableMetric.dump_atexit = False
+        TableMetric.outfile = StringIO()
 
-        list(measure_iter(math_is_hard(10), metric=StatsMetric.metric, name="alice"))
-        list(measure_iter(math_is_hard(20), metric=StatsMetric.metric, name="alice"))
+        list(measure_iter(math_is_hard(10), metric=TableMetric.metric, name="alice"))
+        list(measure_iter(math_is_hard(20), metric=TableMetric.metric, name="alice"))
 
-        list(measure_iter(math_is_hard(10), metric=StatsMetric.metric, name="bob"))
+        list(measure_iter(math_is_hard(10), metric=TableMetric.metric, name="bob"))
 
         # unnamed metrics are dropped
-        list(measure_iter(math_is_hard(10), metric=StatsMetric.metric))
+        list(measure_iter(math_is_hard(10), metric=TableMetric.metric))
 
-        StatsMetric.dump()
+        TableMetric.dump()
         result = 'Name         Count Mean        Count Stddev        Elapsed Mean        Elapsed Stddev        \nalice          15.00               5.00               15.00                 5.00             \nbob            10.00               0.00               10.00                 0.00             \n'
-        self.assertMultiLineEqual(StatsMetric.outfile.getvalue(), result)
+        self.assertMultiLineEqual(TableMetric.outfile.getvalue(), result)
 
 class PlotMetricTestCase(MeasureItTestCase):